pkg/clients: move up non aws specific clients
