+++ cmd/inventory/worker.go
@@ -10,7 +10,7 @@ import (
 	"github.com/hibiken/asynq"
 	"github.com/urfave/cli/v2"
 
-	"github.com/gardener/inventory/pkg/aws/clients"
+	"github.com/gardener/inventory/pkg/clients"
 	"github.com/gardener/inventory/pkg/core/registry"
 )
