+++ pkg/aws/tasks/instances.go
@@ -8,9 +8,10 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/ec2"
 	"github.com/hibiken/asynq"
 
-	"github.com/gardener/inventory/pkg/aws/clients"
+	awsclients "github.com/gardener/inventory/pkg/aws/clients"
 	"github.com/gardener/inventory/pkg/aws/models"
 	"github.com/gardener/inventory/pkg/aws/utils"
+	"github.com/gardener/inventory/pkg/clients"
 	"github.com/gardener/inventory/pkg/utils/strings"
 )
 
@@ -49,7 +50,7 @@ func HandleCollectInstancesRegionTask(ctx context.Context, t *asynq.Task) error 
 func collectInstancesRegion(ctx context.Context, region string) error {
 	slog.Info("Collecting AWS instances ", "region", region)
 
-	instancesOutput, err := clients.Ec2.DescribeInstances(ctx,
+	instancesOutput, err := awsclients.Ec2.DescribeInstances(ctx,
 		&ec2.DescribeInstancesInput{},
 		func(o *ec2.Options) {
 			o.Region = region
