+++ pkg/aws/tasks/vpcs.go
@@ -9,9 +9,10 @@ import (
 	"github.com/aws/aws-sdk-go-v2/service/ec2"
 	"github.com/hibiken/asynq"
 
-	"github.com/gardener/inventory/pkg/aws/clients"
+	awsclients "github.com/gardener/inventory/pkg/aws/clients"
 	"github.com/gardener/inventory/pkg/aws/models"
 	"github.com/gardener/inventory/pkg/aws/utils"
+	"github.com/gardener/inventory/pkg/clients"
 	"github.com/gardener/inventory/pkg/utils/strings"
 )
 
@@ -76,7 +77,7 @@ func HandleCollectVpcsRegionTask(ctx context.Context, t *asynq.Task) error {
 
 func collectVpcsRegion(ctx context.Context, region string) error {
 	slog.Info("Collecting AWS VPCs", "region", region)
-	vpcsOutput, err := clients.Ec2.DescribeVpcs(ctx,
+	vpcsOutput, err := awsclients.Ec2.DescribeVpcs(ctx,
 		&ec2.DescribeVpcsInput{},
 		func(o *ec2.Options) {
 			o.Region = region
